PA Skill calculator moved to rg-stats algorithms module
