+++ typescript/rg-stats/src/index.ts
@@ -12,3 +12,4 @@ export * as MaimaiRate from "./algorithms/maimai-rate";
 export * as Potential from "./algorithms/potential";
 export * as ONGEKIRating from "./algorithms/ongeki-rating";
 export * as DDRFlare from "./algorithms/ddr-flare";
+export * as PolarisChordPaSkill from "./algorithms/poco-paskill";
