+++ typescript/server/src/game-implementations/games/polarischord.ts
@@ -4,6 +4,7 @@ import { CreatePBMergeFor } from "#game-implementations/utils/pb-merge";
 import { ProfileAvgBestN } from "#game-implementations/utils/profile-calc";
 import { SessionAvgBest10For } from "#game-implementations/utils/session-calc";
 import { IsNullish } from "#utils/misc";
+import { PolarisChordSkill } from "rg-stats";
 
 // 100% ACHIEVEMENT RATE is a SSS+ lamp but has its own rank
 const POLARISCHORD_GBOUNDARIES: Array<[string, number]> = [
@@ -29,36 +30,6 @@ function getGrade(score: number): string {
 	return "D";
 }
 
-// PA SKILL formula is weird. Adapted from bemaniwiki with "border" thresholds:
-
-function calculatePASKILL(percent: number, levelNum: number): number {
-	let modifier = 0;
-
-	if (percent >= 100) {
-		modifier = 2.3;
-	} else if (percent >= 99.5) {
-		modifier = 2.0 + (percent - 99.5) * 0.6;
-	} else if (percent >= 98.0) {
-		modifier = 0.5 + (percent - 98.0) * 1.0;
-	} else if (percent >= 95.0) {
-		modifier = 0.0 + (percent - 95.0) * (0.5 / 3.0);
-	} else if (percent >= 85.0) {
-		modifier = -1.0 + (percent - 85.0) * 0.1;
-	} else if (percent >= 80.0) {
-		const at80 = -(levelNum + 3) / 4.0;
-		const slope = (-1.0 - at80) / 5.0;
-		modifier = at80 + (percent - 80.0) * slope;
-	} else if (percent >= 50.0) {
-		const at50 = -levelNum;
-		const at80 = -(levelNum + 3) / 4.0;
-		const slope = (at80 - at50) / 30.0;
-		modifier = at50 + (percent - 50.0) * slope;
-	} else {
-		modifier = -levelNum;
-	}
-
-	return Number((levelNum + modifier).toFixed(5));
-}
 
 export const POLARISCHORD_IMPL: GameImplementation<"polarischord"> = {
 	chartSpecificValidators: {},
@@ -66,7 +37,7 @@ export const POLARISCHORD_IMPL: GameImplementation<"polarischord"> = {
 		grade: getGrade(scoreData.percent),
 	}),
 	scoreCalcs: (scoreData, _derivedData, chart) => ({
-		paSkill: calculatePASKILL(scoreData.percent, chart.levelNum),
+		paSkill: PolarisChordPaSkill.calculate(scoreData.percent, chart.levelNum),
 	}),
 	pbRankingValues: (pb) => ({
 		ranking: pb.scoreData.percent,
